Use picoresolve to resolve javadoc artifacts
Also now compiling graal native images with x86-64-v4 since that is
what our little server is running on. Also limited the amount of memory
the remote process can use. 64 megs on memory should be sufficent,
but one can always upgrade that number if need be.

Outside of that there is a tiny bug in this commit where
any string will be interpreted as being the latest version. I have not
yet bothered to figure that one out as "standard" version ranges seem
to work and cause a 404. And generally a false 200 is always better
than a false 404, so this is no serious issue.