Allow to exclude artifacts from javadoc index
This is currently only used to hide artifacts that are now being published
under different coordinates from the artifact list view. However, in theory
old and outdated artifacts can be hidden with this feature, too.

Note that hidden artifacts only are excluded from the central javadoc index,
but are still present in the nexus maven repository index and the javadoc
contents are still queryable as has been the case previously using a direct
link. Further, version ranges in such links are not impacted by this commit.