Search API: Initial chunk of changes
Currently does not contain any behavioural changes. This is mostly caused by the fact that
the search API depends on a smaller rework of how updates to the repository are propagated.

Also, the actual frontend is not yet developed (nor do I have a real idea on how the frontend
should look like, but I suppose I'll figure that out later on)