Flush pipeline
This commit mainly contains work that has been left uncompleted since
October 2025. It's likely that I'll do a tabula rasa on that work though,
since evidently it's not in a good shape if it has been left untouched
for 10 months.

Though I guess for the standards of many projects that isn't actually
too long.